+++ src/components/Details.styles.js
