+++ src/components/app-bar/SearchBox.styles.js
