+++ src/components/auth/AuthModal.styles.js
