+++ src/firebase/firebase-config.js
