+++ src/lib/aux-functions.js
