+++ src/lib/youtube-api.js
@@ -20,8 +20,8 @@ const fetchData = async (params) => {
 
 const searchVideos = async (params) => {
   const data = await fetchData(params);
-  console.log('search videos');
-  console.log(data);
+  // console.log('search videos');
+  // console.log(data);
 
   const filteredData = data.items.map((video) => {
     const {
@@ -46,8 +46,8 @@ const searchVideos = async (params) => {
 
 const searchVideoDurations = async (videos, params) => {
   const data = await fetchData(params);
-  console.log('search video durations');
-  console.log(data);
+  // console.log('search video durations');
+  // console.log(data);
 
   return videos.map((video) => {
     return {
@@ -60,8 +60,8 @@ const searchVideoDurations = async (videos, params) => {
 
 const searchChannelLogos = async (videos, params) => {
   const data = await fetchData(params);
-  console.log('search channel logos');
-  console.log(data);
+  // console.log('search channel logos');
+  // console.log(data);
 
   return videos.map((video) => {
     const channelLogo =
